Add theme font options to the editor Settings
The text in the editor is displayed in the Arial font by default. This commit will add 2 x 2 extra options so it may be changed to the user preferences.

The theme is at its best with the Open Sans font so that text fields aren't cut off, because the width of the Arial characters is generally large. However, Arial remains to be default, also the fallback.

Console/engine function is added to be able to read environment variables in script; In this case to be able to find the fonts installation path.