Switch default to 'Segoe UI'  (Win sys font) from Arial
Switch to the Segoe UI windows system font instead of Arial for the GUI's. This will solve the problem of overlapping GUI elements and in most instances text being cut off (labels etc.)

Segoe UI is supported/provided on/with Windows Vista up to Windows 11 and especially designed for use in GUI's

This commit does not tackle the security issues with the "getEnvironment" command.